COMMERCE-1950 talend - component - Reorganize - abandon classes and introduce SchemaBuilder
